pyproject.toml
[project]
name = "are"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "db-dtypes>=1.4.3",
    "google-cloud-bigquery-storage>=2.33.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "vertexai>=1.71.1",
    "fastapi[standard]>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "google-cloud-monitoring>=2.15.0",
    "google-cloud-logging>=3.8.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
]
